+++ backend/src/main/java/com/woowacourse/ternoko/auth/presentation/AuthInterceptor.java
@@ -4,7 +4,7 @@ import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
 import com.woowacourse.ternoko.auth.application.AuthService;
 import com.woowacourse.ternoko.common.exception.ExceptionType;
-import com.woowacourse.ternoko.common.exception.exception.TokenInvalidException;
+import com.woowacourse.ternoko.common.exception.TokenInvalidException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
