--- backend/src/main/java/com/woowacourse/ternoko/common/exception/InvalidMemberNicknameException.java
@@ -1,10 +1,0 @@
-package com.woowacourse.ternoko.common.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class InvalidMemberNicknameException extends CommonException {
-
-    public InvalidMemberNicknameException(final ExceptionType exceptionType) {
-        super(HttpStatus.BAD_REQUEST, exceptionType.getStatusCode(), exceptionType.getMessage());
-    }
-}
