+++ backend/src/main/java/com/woowacourse/ternoko/common/exception/TokenInvalidException.java
@@ -1,6 +1,5 @@
-package com.woowacourse.ternoko.common.exception.exception;
+package com.woowacourse.ternoko.common.exception;
 
-import com.woowacourse.ternoko.common.exception.ExceptionType;
 import com.woowacourse.ternoko.common.exception.advice.CommonException;
 
 public class TokenInvalidException extends CommonException {
