+++ backend/src/main/java/com/woowacourse/ternoko/core/application/CrewService.java
@@ -2,11 +2,9 @@ package com.woowacourse.ternoko.core.application;
 
 import static com.woowacourse.ternoko.common.exception.ExceptionType.DUPLICATED_MEMBER_NICKNAME;
 
-import com.woowacourse.ternoko.common.exception.CrewNotFoundException;
+import com.woowacourse.ternoko.common.exception.CrewInvalidException;
 import com.woowacourse.ternoko.common.exception.ExceptionType;
-import com.woowacourse.ternoko.common.exception.InvalidMemberNicknameException;
 import com.woowacourse.ternoko.core.domain.member.MemberRepository;
-import com.woowacourse.ternoko.common.exception.exception.CrewInvalidException;
 import com.woowacourse.ternoko.core.domain.member.crew.Crew;
 import com.woowacourse.ternoko.core.domain.member.crew.CrewRepository;
 import com.woowacourse.ternoko.core.dto.request.CrewUpdateRequest;
@@ -29,7 +27,7 @@ public class CrewService {
         final String requestNickname = request.getNickname();
 
         if (memberRepository.existsByIdAndNicknameExceptMe(crewId, requestNickname)) {
-            throw new InvalidMemberNicknameException(DUPLICATED_MEMBER_NICKNAME);
+            throw new CrewInvalidException(DUPLICATED_MEMBER_NICKNAME, crewId);
         }
         crewRepository.updateNickNameAndImageUrl(crewId, request.getNickname(), request.getImageUrl());
     }
