+++ backend/src/test/java/com/woowacourse/ternoko/core/domain/interview/formitem/QuestionTest.java
@@ -4,7 +4,7 @@ import static com.woowacourse.ternoko.common.exception.ExceptionType.OVER_LENGTH
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import com.woowacourse.ternoko.common.exception.exception.InterviewInvalidException;
+import com.woowacourse.ternoko.common.exception.InterviewInvalidException;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
